+++ crates/containerd-shim-wasm/src/sandbox/instance.rs
@@ -6,8 +6,6 @@ use std::time::Duration;
 use chrono::{DateTime, Utc};
 
 use super::error::Error;
-use super::sync::WaitableCell;
-use crate::sys::signals::*;
 
 /// Generic options builder for creating a wasm instance.
 /// This is passed to the `Instance::new` method.
@@ -141,93 +139,8 @@ pub trait Instance: 'static {
         self.wait_timeout(None).unwrap()
     }
 
-    /// Waits for the instance to finish and retunrs its exit code
+    /// Waits for the instance to finish and returns its exit code
     /// Returns None if the timeout is reached before the instance has finished.
     /// This is a blocking call.
     fn wait_timeout(&self, t: impl Into<Option<Duration>>) -> Option<(u32, DateTime<Utc>)>;
-}
-
-/// This is used for the "pause" container with cri and is a no-op instance implementation.
-pub struct Nop {
-    /// Since we are faking the container, we need to keep track of the "exit" code/time
-    /// We'll just mark it as exited when kill is called.
-    exit_code: WaitableCell<(u32, DateTime<Utc>)>,
-}
-
-impl Instance for Nop {
-    type Engine = ();
-    fn new(_id: String, _cfg: Option<&InstanceConfig<Self::Engine>>) -> Result<Self, Error> {
-        Ok(Nop {
-            exit_code: WaitableCell::new(),
-        })
-    }
-    fn start(&self) -> Result<u32, Error> {
-        Ok(std::process::id())
-    }
-    fn kill(&self, signal: u32) -> Result<(), Error> {
-        let code = match signal as i32 {
-            SIGKILL => 137,
-            SIGINT | SIGTERM => 0,
-            s => {
-                return Err(Error::InvalidArgument(format!("unsupported signal: {}", s)));
-            }
-        };
-
-        let _ = self.exit_code.set((code, Utc::now()));
-
-        Ok(())
-    }
-    fn delete(&self) -> Result<(), Error> {
-        Ok(())
-    }
-    fn wait_timeout(&self, t: impl Into<Option<Duration>>) -> Option<(u32, DateTime<Utc>)> {
-        self.exit_code.wait_timeout(t).copied()
-    }
-}
-
-#[cfg(test)]
-mod noptests {
-    use std::time::Duration;
-
-    use super::*;
-
-    #[test]
-    fn test_nop_kill_sigkill() -> Result<(), Error> {
-        let nop = Nop::new("".to_string(), None)?;
-
-        nop.kill(SIGKILL as u32)?;
-
-        let ec = nop.wait_timeout(Duration::from_secs(3)).unwrap();
-        assert_eq!(ec.0, 137);
-        Ok(())
-    }
-
-    #[test]
-    fn test_nop_kill_sigterm() -> Result<(), Error> {
-        let nop = Nop::new("".to_string(), None)?;
-
-        nop.kill(SIGTERM as u32)?;
-
-        let ec = nop.wait_timeout(Duration::from_secs(3)).unwrap();
-        assert_eq!(ec.0, 0);
-        Ok(())
-    }
-
-    #[test]
-    fn test_nop_kill_sigint() -> Result<(), Error> {
-        let nop = Nop::new("".to_string(), None)?;
-
-        nop.kill(SIGINT as u32)?;
-
-        let ec = nop.wait_timeout(Duration::from_secs(3)).unwrap();
-        assert_eq!(ec.0, 0);
-        Ok(())
-    }
-
-    #[test]
-    fn test_nop_delete_after_create() -> Result<(), Error> {
-        let nop = Nop::new("".to_string(), None)?;
-        nop.delete()?;
-        Ok(())
-    }
 }