+++ crates/containerd-shim-wasm/src/sandbox/shim/mod.rs
@@ -5,7 +5,6 @@
 mod cli;
 mod events;
 mod instance_data;
-mod instance_option;
 mod local;
 #[cfg(feature = "opentelemetry")]
 mod otel;
@@ -13,4 +12,4 @@ mod task_state;
 
 pub use cli::Cli;
 #[cfg(feature = "opentelemetry")]
-pub use otel::{traces_enabled as otel_traces_enabled, Config as OTLPConfig};
+pub use otel::{traces_enabled as otel_traces_enabled, Config as OtlpConfig};