+++ crates/containerd-shim-wasm/src/sys/unix/signals.rs
@@ -1,1 +1,0 @@
-pub use libc::{SIGINT, SIGKILL, SIGTERM};