+++ crates/containerd-shim-wasm/src/sys/windows/container/instance.rs
@@ -33,7 +33,7 @@ impl<E: Engine> SandboxInstance for Instance<E> {
         todo!();
     }
 
-    /// Waits for the instance to finish and retunrs its exit code
+    /// Waits for the instance to finish and returns its exit code
     /// Returns None if the timeout is reached before the instance has finished.
     /// This is a blocking call.
     fn wait_timeout(&self, _t: impl Into<Option<Duration>>) -> Option<(u32, DateTime<Utc>)> {